+++ images/03.ServiceImages/dns/home/engines/scripts/signal/_kill_bind.sh
@@ -1,19 +1,18 @@
 #!/bin/sh
+if test -f /var/run/named/named.pid
+ then
+  pid=`cat /var/run/named/named.pid`
 
-pid=`cat /var/run/named/named.pid`
-#if ! test $1 = HUP
-# then 
-#  rm /var/run/named/named.pid
-#fi
-# 
-kill -$1 $pid
-
-
-
-if ! test $1 = HUP
- then 
-if test -f  /var/run/named/named.pid
-  rm /var/run/named/named.pid
-  fi
+  if test `kill -0 $pid &>/dev/null` -eq 0
+   then
+     kill -$1 $pid
+   fi
+   
+    if ! test $1 = HUP
+     then 
+      if test -f  /var/run/named/named.pid
+       then
+        rm /var/run/named/named.pid
+      fi
+    fi
 fi
-
